Feels like we're actually getting somewhere.
We can check if we have the items in our inventory that are necessary. If we have them, we can take care of the job. If we don't we can locate those items in stockpiles.

We need to be able to remove the item(s) from our inventory if we already have what it costs. Working that into the bigger inventory shifts coming up.